Hash full offer data consistently
We need to hash all offer params so clients can verify offer
given only the offer data and offerHash, without looking at the blockchain.

We also need to ensure integers are converted to bytes consistently.
To do that we pad the amounts with 0s up to 32 bytes.
This effectively limits the amounts to 32 bytes but it should be
sufficient, as it is the same as ethereum's largest uint256.